Split out keyboard pieces and factor to subfolder, add resolution strategy
